+++ pennylane_lightning/lightning_qubit/lightning_qubit.py
@@ -42,7 +42,7 @@ from pennylane.operation import DecompositionUndefinedError, Operator
 from pennylane.ops import Conditional, PauliRot, Prod, SProd, Sum
 from pennylane.tape import QuantumScript
 from pennylane.transforms.core import TransformProgram
-from pennylane.typing import Result, TensorLike
+from pennylane.typing import Result
 
 from pennylane_lightning.core.lightning_base import (
     LightningBase,
@@ -528,90 +528,6 @@ class LightningQubit(LightningBase):
                 return "LightningSimulator", lib_location
 
         raise RuntimeError("'LightningSimulator' shared library not found")  # pragma: no cover
-
-    def jaxpr_jvp(
-        self,
-        jaxpr: "jax.core.Jaxpr",
-        args: Sequence[TensorLike],
-        tangents: Sequence[TensorLike],
-        execution_config: Optional[ExecutionConfig] = None,
-    ) -> tuple[Sequence[TensorLike], Sequence[TensorLike]]:
-        """
-        An **experimental** method for computing the results and jvp for PLXPR with LightningQubit.
-
-        Args:
-            jaxpr (jax.core.Jaxpr): Pennylane variant jaxpr containing quantum operations
-                and measurements
-            args (Sequence[TensorLike]): the ``consts`` followed by the normal arguments
-            tangents (Sequence[TensorLike]): the tangents corresponding to ``args``.
-                May contain ``jax.interpreters.ad.Zero``.
-
-        Keyword Args:
-            execution_config (Optional[ExecutionConfig]): a data structure with additional information required for execution
-
-        Returns:
-            Sequence[TensorLike], Sequence[TensorLike]: the results and jacobian vector products
-
-
-        .. note::
-
-            For LightningQubit, the current implementation of this method is based on the conversion of the jaxpr to a PennyLane tape.
-            This has strict limitations. The ``args`` should contain the concatenation of ``jaxpr.constvars`` and ``jaxpr.invars``,
-            which are assumed to represent the trainable parameters of the circuit.
-            The method will raise an error if ``args`` do not match exactly the parameters of the jaxpr converted to quantum tape.
-
-        """
-
-        # pylint: disable=import-outside-toplevel
-        import jax
-
-        from pennylane_lightning.core.jaxpr_adjoint import (
-            convert_jaxpr_to_tape,
-            get_output_shapes,
-            validate_args_tangents,
-        )
-
-        if self.wires is None:
-            raise NotImplementedError("Wires must be specified for integration with plxpr capture.")
-
-        gradient_method = getattr(execution_config, "gradient_method", "adjoint")
-
-        if gradient_method != "adjoint":
-            raise NotImplementedError(
-                f"LightningQubit does not support gradient_method={gradient_method} for jaxpr_jvp."
-            )
-
-        tangents = validate_args_tangents(args, tangents)
-
-        self._statevector = self.LightningStateVector(
-            num_wires=len(self.wires), dtype=self._c_dtype
-        )
-
-        def wrapper(*args):
-            """
-            Evaluate a jaxpr by converting it into a quantum tape, ensuring the provided
-            parameters match the tape's parameters, and then simulating the tape to compute
-            both the result and the Jacobian.
-
-            The *args should contain the concatenation of jaxpr.constvars and jaxpr.invars,
-            which are assumed to represent the trainable parameters.
-            """
-            tape = convert_jaxpr_to_tape(jaxpr, args)
-            return self.simulate_and_jacobian(tape, state=self._statevector)
-
-        shapes_res, shapes_jac = get_output_shapes(jaxpr, self.shots, len(self.wires))
-        results, jacobians = jax.pure_callback(wrapper, (shapes_res, shapes_jac), *args)
-
-        if len(tangents) == 1 and not jax.numpy.isscalar(tangents[0]):
-            tangents = tangents[0]
-
-        jvps = (
-            [qml.gradients.compute_jvp_single(tangents, jacobians)]
-            if len(jaxpr.outvars) == 1
-            else qml.gradients.compute_jvp_multi(tangents, jacobians)
-        )
-
-        return results, jvps
 
 
 _supports_operation = LightningQubit.capabilities.supports_operation
